# docs/api/openapi.yaml
openapi: 3.0.3

info:
  title: Expense Tracker API
  version: 0.1.0
  license:
    name: MIT
  description: |
    REST API для expense tracker.

    ## Соглашения

    - **Базовый путь:** `/api`
    - **Формат дат:** ISO 8601 (`2026-07-13T10:30:00Z`).
    - **Денёжные суммы:** `integer (int64)` в минорных единицах. $12.50 → `1250`.
    - **ID:** UUID v4 строки.
    - **Content-Type:** `application/json` для всех request bodies.
    - **Auth:** stateful session cookie (`session_id`).

    ## Idempotency

    `POST /api/transactions` поддерживает идемпотентность через заголовок
    `Idempotency-Key`. Повторный запрос с тем же ключом и тем же body
    возвращает закешированный ответ. Повтор с тем же ключом и другим body →
    `409 Conflict`.

    ## Optimistic Concurrency для PATCH

    `PATCH /api/transactions/{id}` поддерживает optimistic concurrency через
    обязательное поле `version` в теле запроса. Клиент получает `version` из
    `GET /api/transactions/{id}` или из ответа на `POST /api/transactions` и
    отправляет его обратно при PATCH. Если запись была изменена параллельно —
    сервер возвращает `409 Conflict` с кодом `TRANSACTION_VERSION_CONFLICT`.

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Auth
    description: Регистрация, вход, выход, текущий пользователь
  - name: Transactions
    description: Операции с транзакциями (income / expense / transfer)

security:
  - sessionCookie: []

paths:
  /api/auth/register:
    post:
      tags: [Auth]
      operationId: registerUser
      summary: Регистрация пользователя
      description: |
        Создаёт пользователя и сразу стартует сессию (Set-Cookie).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 8
                  maxLength: 72
                  description: Plaintext (хешируется на сервере).
                  example: strong-password
      responses:
        "201":
          description: Пользователь создан, сессия установлена.
          headers:
            Set-Cookie:
              description: session_id cookie.
              schema:
                type: string
                example: session_id=abc; Path=/; HttpOnly
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          description: Email уже занят.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: USER_ALREADY_EXISTS
                message: user already exists
        "500":
          $ref: "#/components/responses/InternalError"

  /api/auth/login:
    post:
      tags: [Auth]
      operationId: loginUser
      summary: Вход
      description: |
        Проверяет credentials, стартует сессию. Rate-limited по ClientIP
        (по умолчанию блокировка после 5 неудачных попыток в течение минуты).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Успешный вход, сессия установлена.
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          description: Неверный email или пароль.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: INVALID_CREDENTIALS
                message: invalid credentials
        "429":
          description: Слишком неудачных попыток входа с этого IP.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                code: TOO_MANY_REQUESTS
                message: too many login attempts, please try again later

  /api/auth/logout:
    post:
      tags: [Auth]
      operationId: logoutUser
      summary: Выход
      description: |
        Удаляет текущую сессию на сервере и сбрасывает cookie. Если cookie
        отсутствует — молча возвращает 204.
      responses:
        "204":
          description: Успешный выход (или cookie уже отсутствовал).
          headers:
            Set-Cookie:
              description: session_id cookie с Max-Age=-1 (browser удалит).
              schema:
                type: string
        "500":
          $ref: "#/components/responses/InternalError"

  /api/auth/me:
    get:
      tags: [Auth]
      operationId: getCurrentUser
      summary: Текущий пользователь
      responses:
        "200":
          description: Текущий пользователь по session cookie.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /api/transactions:
    get:
      tags: [Transactions]
      operationId: listTransactions
      summary: Список транзакций текущего пользователя
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense, transfer]
        - name: accountId
          in: query
          schema:
            type: string
            format: uuid
        - name: categoryId
          in: query
          schema:
            type: string
            format: uuid
        - name: fromDate
          in: query
          description: ISO date (YYYY-MM-DD), lower bound inclusive.
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          description: ISO date (YYYY-MM-DD), upper bound inclusive.
          schema:
            type: string
            format: date
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
        - name: sort
          in: query
          schema:
            type: string
            enum: [occurredAt, -occurredAt, amount, -amount]
            default: occurredAt
      responses:
        "200":
          description: Список транзакций (опционально отфильтрованный).
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [Transactions]
      operationId: createTransaction
      summary: Создать транзакцию
      description: |
        Поддерживает идемпотентность. Если заголовок `Idempotency-Key`
        присутствует, повторный запрос с тем же ключом + тем же body
        возвращает закешированный ответ. Тот же ключ + другой body → `409`.
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionCreateRequest"
            examples:
              cashflow:
                summary: income / expense
                value:
                  type: income
                  amount: 100000
                  description: Salary for June
                  occurredAt: "2026-07-13T10:30:00Z"
                  accountId: 550e8400-e29b-41d4-a716-446655440000
                  categoryId: 550e8400-e29b-41d4-a716-446655440001
              transfer:
                summary: transfer between accounts
                value:
                  type: transfer
                  amount: 50000
                  description: To savings
                  occurredAt: "2026-07-13T10:30:00Z"
                  fromAccountId: 550e8400-e29b-41d4-a716-446655440002
                  toAccountId: 550e8400-e29b-41d4-a716-446655440003
      responses:
        "201":
          description: Транзакция создана.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: |
            Idempotency конфликт:
            - тот же ключ + другой body (mismatch)
            - тот же ключ пока предыдущий запрос ещё в полёте (pending)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                mismatch:
                  value:
                    code: IDEMPOTENCY_KEY_MISMATCH
                    message: idempotency key request hash mismatch
                inUse:
                  value:
                    code: IDEMPOTENCY_KEY_IN_USE
                    message: idempotency key already used
        "422":
          description: |
            Бизнес-правило нарушено — ссылки не найдены или type-mismatch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                accountNotFound:
                  value:
                    code: ACCOUNT_NOT_FOUND
                    message: account not found
                categoryNotFound:
                  value:
                    code: CATEGORY_NOT_FOUND
                    message: category not found
                categoryTypeMismatch:
                  value:
                    code: CATEGORY_TYPE_MISMATCH
                    message: transaction type does not match category type
                sameAccountTransfer:
                  value:
                    code: SAME_ACCOUNT_TRANSFER
                    message: transaction from and to accounts are the same
                invalidRefs:
                  value:
                    code: INVALID_REFS
                    message: invalid references
        "500":
          $ref: "#/components/responses/InternalError"

  /api/transactions/{id}:
    parameters:
      - $ref: "#/components/parameters/TransactionId"
    get:
      tags: [Transactions]
      operationId: getTransaction
      summary: Получить транзакцию по ID
      responses:
        "200":
          description: Транзакция.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/TransactionNotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    patch:
      tags: [Transactions]
      operationId: updateTransaction
      summary: Частично обновить транзакцию
      description: |
        Все поля optional. Поле `type` менять нельзя — оно вычисляется из
        набора ссылок (cashflow vs transfer).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionUpdateRequest"
      responses:
        "200":
          description: Обновлённая транзакция.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: |
            Либо validation failed, либо пустой body (`no fields to update`).
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/ErrorResponse"
                  - $ref: "#/components/schemas/ValidationErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/TransactionNotFound"
        "409":
          $ref: "#/components/responses/TransactionVersionConflict"
        "422":
          description: То же, что у POST (бизнес-правило нарушено).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      tags: [Transactions]
      operationId: deleteTransaction
      summary: Удалить транзакцию
      responses:
        "204":
          description: Удалено.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/TransactionNotFound"
        "500":
          $ref: "#/components/responses/InternalError"

components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session_id

  parameters:
    TransactionId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Уникальный client-generated ключ (UUID recommended). Один ключ =
        один intent. Повтор того же запроса возвращает тот же ответ.
      schema:
        type: string
        maxLength: 255

  responses:
    Unauthorized:
      description: Не авторизован (нет валидной session cookie).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: UNAUTHORIZED
            message: missing session cookie
    InternalError:
      description: Внутренняя ошибка сервера.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: INTERNAL_ERROR
            message: internal server error
    ValidationError:
      description: Неверное тело запроса (validation failed).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationErrorResponse"
          example:
            code: VALIDATION_FAILED
            message: validation failed
            errors:
              - field: email
                message: email is required
    TransactionNotFound:
      description: Транзакция не найдена (или чужая).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: TRANSACTION_NOT_FOUND
            message: transaction not found
    TransactionVersionConflict:
      description: Конфликт в версии транзакции (optimistic concurrency).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: TRANSACTION_VERSION_CONFLICT
            message: transaction version conflict

  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, email, createdAt, updatedAt]

    Transaction:
      type: object
      description: |
        Cashflow-транзакция (income/expense) содержит `accountId`+`categoryId`.
        Transfer содержит `fromAccountId`+`toAccountId`. Эти пары
        взаимоисключающие.
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        type:
          type: string
          enum: [income, expense, transfer]
        amount:
          type: integer
          format: int64
          description: Минорные единицы (divisor 100 для USD/EUR/RUB).
        description:
          type: string
        occurredAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        accountId:
          type: string
          format: uuid
          nullable: true
        categoryId:
          type: string
          format: uuid
          nullable: true
        fromAccountId:
          type: string
          format: uuid
          nullable: true
        toAccountId:
          type: string
          format: uuid
          nullable: true
        version:
          type: integer
          format: int
          description: Версия транзакции
      required:
        [
          id,
          userId,
          type,
          amount,
          description,
          occurredAt,
          createdAt,
          updatedAt,
          version,
        ]

    TransactionCreateRequest:
      type: object
      required: [type, amount, occurredAt]
      properties:
        type:
          type: string
          enum: [income, expense, transfer]
        amount:
          type: integer
          format: int64
          minimum: 1
        description:
          type: string
          default: ""
        occurredAt:
          type: string
          format: date-time
        accountId:
          type: string
          format: uuid
          description: Required для income/expense. Forbidden для transfer.
        categoryId:
          type: string
          format: uuid
          description: Required для income/expense. Forbidden для transfer.
        fromAccountId:
          type: string
          format: uuid
          description: Required для transfer. Forbidden для income/expense.
        toAccountId:
          type: string
          format: uuid
          description: Required для transfer. Forbidden для income/expense.

    TransactionUpdateRequest:
      type: object
      description: Все поля кроме `version` optional. Поле `type` менять нельзя.
      properties:
        version:
          type: integer
          format: int
          minimum: 1
        amount:
          type: integer
          format: int64
          minimum: 1
        description:
          type: string
        occurredAt:
          type: string
          format: date-time
        accountId:
          type: string
          format: uuid
        categoryId:
          type: string
          format: uuid
        fromAccountId:
          type: string
          format: uuid
        toAccountId:
          type: string
          format: uuid
      required: [version]

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Machine-readable error code.
          example: VALIDATION_FAILED
        message:
          type: string
          description: Human-readable сообщение.
      required: [code, message]

    ValidationErrorResponse:
      allOf:
        - $ref: "#/components/schemas/ErrorResponse"
        - type: object
          properties:
            errors:
              type: array
              items:
                $ref: "#/components/schemas/FieldError"

    FieldError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
      required: [field, message]